requirements.txt
fastapi[standard]==0.116.1
uvicorn==0.32.0
SQLAlchemy==2.0.43
python-dotenv==1.1.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
lxml==6.0.1
boto3==1.35.36
python-multipart==0.0.12

google-api-python-client==2.182.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.38.0
pydantic-settings==2.10.1
